cpp_black_belt (week_5): Part T size_t -> double

+++ cpp_black_belt/transport_final_project/transport_catalog/time_interval.cpp
@@ -6,11 +6,11 @@ std::optional<std::pair<It,It>> FindDayIntervals(Time::DayT day, It intervals_be
 {
     auto it_lower = std::lower_bound(intervals_begin, intervals_end, day,
     [](const TimeInterval& interval, Time::DayT day_){
-        return static_cast<size_t>(interval.Day()) < static_cast<size_t>(day_);
+        return static_cast<double>(interval.Day()) < static_cast<double>(day_);
     });
     auto it_upper = std::upper_bound(intervals_begin, intervals_end, day,
     [](Time::DayT day_, const TimeInterval& interval){
-        return static_cast<size_t>(day_) < static_cast<size_t>(interval.Day());
+        return static_cast<double>(day_) < static_cast<double>(interval.Day());
     });
     
     for(auto it = it_lower; it != it_upper; it++) {
@@ -25,9 +25,9 @@ std::optional<std::pair<It,It>> FindDayIntervals(Time::DayT day, It intervals_be
 };
 
 template<typename It>
-std::optional<size_t> ProcessDayIntervals (It it_begin, It it_end, size_t mins) {
+std::optional<double> ProcessDayIntervals (It it_begin, It it_end, double mins) {
     auto it = std::upper_bound(it_begin, it_end, mins,
-    [](size_t time, const TimeInterval& interval){
+    [](double time, const TimeInterval& interval){
         return time < interval.MinsTo();
     });
     if(it == it_end) {
@@ -38,7 +38,7 @@ std::optional<size_t> ProcessDayIntervals (It it_begin, It it_end, size_t mins) 
     return 0ul;
 };
 
-size_t GetWaitingTime(const Time& start_time, const std::vector<TimeInterval>& intervals) {
+double GetWaitingTime(const Time& start_time, const std::vector<TimeInterval>& intervals) {
     if(intervals.empty()) {
         return 0;
     }
@@ -51,8 +51,8 @@ size_t GetWaitingTime(const Time& start_time, const std::vector<TimeInterval>& i
             return *ProcessDayIntervals(intervals.begin(),intervals.end(),0ul) + 24 * 60 - start_time.Mins();
         }
     } else {
-        size_t cur_mins = start_time.Mins();
-        size_t acced_mins = 0;
+        double cur_mins = start_time.Mins();
+        double acced_mins = 0;
         for(auto day = start_time.Day();;day = Time::IncDay(day)) {
             if(auto opt_it = FindDayIntervals(day,intervals.begin(),intervals.end()); opt_it) {
                 auto [it_begin,it_end] = *opt_it;
